+++ src/main/java/com/knu/sosuso/capstone/domain/comment/dto/CommentDto.java
@@ -1,9 +1,11 @@
-package com.knu.sosuso.capstone.domain.conmment.dto;
+package com.knu.sosuso.capstone.domain.comment.dto;
 
-public record ReplyDto(
+public record CommentDto(
         String id,
         String author,
         String text,
         Integer likeCount,
+        String sentiment,
         String publishedAt
-) {}
+) {
+}
