+++ src/main/java/com/knu/sosuso/capstone/domain/comment/service/CommentQueryService.java
@@ -1,11 +1,11 @@
-package com.knu.sosuso.capstone.domain.conmment.service;
+package com.knu.sosuso.capstone.domain.comment.service;
 
-import com.knu.sosuso.capstone.domain.conmment.entity.Comment;
+import com.knu.sosuso.capstone.domain.comment.entity.Comment;
 import com.knu.sosuso.capstone.domain.video.entity.Video;
-import com.knu.sosuso.capstone.domain.conmment.entity.value.SentimentType;
-import com.knu.sosuso.capstone.domain.conmment.dto.CommentDto;
-import com.knu.sosuso.capstone.domain.conmment.dto.response.CommentResponse;
-import com.knu.sosuso.capstone.domain.conmment.repository.CommentRepository;
+import com.knu.sosuso.capstone.domain.comment.entity.value.SentimentType;
+import com.knu.sosuso.capstone.domain.comment.dto.CommentDto;
+import com.knu.sosuso.capstone.domain.comment.dto.response.CommentResponse;
+import com.knu.sosuso.capstone.domain.comment.repository.CommentRepository;
 import com.knu.sosuso.capstone.domain.video.repository.VideoRepository;
 import com.knu.sosuso.capstone.global.exception.BusinessException;
 import com.knu.sosuso.capstone.global.exception.error.CommentError;
