--- src/main/java/com/knu/sosuso/capstone/domain/conmment/dto/response/CommentResponse.java
@@ -1,11 +1,0 @@
-package com.knu.sosuso.capstone.domain.conmment.dto.response;
-
-import com.knu.sosuso.capstone.domain.conmment.dto.CommentDto;
-
-import java.util.List;
-
-public record CommentResponse(
-        String apiVideoId,
-        List<CommentDto> results
-) {
-}
