+++ src/main/java/com/knu/sosuso/capstone/domain/video/service/VideoProcessingService.java
@@ -3,13 +3,13 @@ package com.knu.sosuso.capstone.domain.video.service;
 import com.knu.sosuso.capstone.domain.ai.dto.AIAnalysisRequest;
 import com.knu.sosuso.capstone.domain.ai.dto.AIAnalysisResponse;
 import com.knu.sosuso.capstone.domain.ai.service.AnalysisService;
-import com.knu.sosuso.capstone.domain.conmment.entity.Comment;
-import com.knu.sosuso.capstone.domain.conmment.service.CommentService;
-import com.knu.sosuso.capstone.domain.conmment.dto.response.CommentApiResponse;
+import com.knu.sosuso.capstone.domain.comment.entity.Comment;
+import com.knu.sosuso.capstone.domain.comment.service.CommentService;
+import com.knu.sosuso.capstone.domain.comment.dto.response.CommentApiResponse;
 import com.knu.sosuso.capstone.domain.detail.dto.DetailChannelDto;
 import com.knu.sosuso.capstone.domain.detail.dto.DetailPageResponse;
 import com.knu.sosuso.capstone.domain.detail.dto.DetailVideoDto;
-import com.knu.sosuso.capstone.domain.conmment.repository.CommentRepository;
+import com.knu.sosuso.capstone.domain.comment.repository.CommentRepository;
 import com.knu.sosuso.capstone.domain.video.entity.Video;
 import com.knu.sosuso.capstone.domain.video.dto.response.VideoApiResponse;
 import com.knu.sosuso.capstone.domain.video.repository.VideoRepository;
