+++ src/main/java/com/knu/sosuso/capstone/global/swagger/ReplyControllerSwagger.java
@@ -1,6 +1,6 @@
 package com.knu.sosuso.capstone.global.swagger;
 
-import com.knu.sosuso.capstone.domain.conmment.dto.response.ReplyResponse;
+import com.knu.sosuso.capstone.domain.comment.dto.response.ReplyResponse;
 import com.knu.sosuso.capstone.global.ResponseDto;
 import com.knu.sosuso.capstone.global.swagger.annotation.ErrorCode400;
 import com.knu.sosuso.capstone.global.swagger.annotation.ErrorCode500;
